refactor: MineMap 생성 방식을 정적 팩토리 메서드 -> 팩토리 패턴 적용
