+++ src/main/kotlin/domain/map/MineMapFactory.kt
@@ -1,0 +1,8 @@
+package domain.map
+
+import domain.MineSweeperInitProperty
+
+interface MineMapFactory {
+
+    fun create(mineSweeperInitProperty: MineSweeperInitProperty): MineMap
+}
