+++ src/main/kotlin/view/MineSweeperController.kt
@@ -1,20 +1,16 @@
 package view
 
-import domain.map.MineMap
-import domain.mine.MineCoordinatesCreator
+import domain.map.MineMapFactory
 
 class MineSweeperController(
     private val inputView: MineSweeperInputView,
     private val resultView: MineSweeperResultView,
-    private val mineCoordinatesCreator: MineCoordinatesCreator,
+    private val mineMapFactory: MineMapFactory,
 ) {
 
     fun start() {
         val mineSweeperInitProperty = inputView.readInitProperty()
-        val mineMap = MineMap.create(
-            mineSweeperInitProperty = mineSweeperInitProperty,
-            mineCoordinatesCreator = mineCoordinatesCreator,
-        )
+        val mineMap = mineMapFactory.create(mineSweeperInitProperty)
         resultView.display(mineMap)
     }
 }
